Add files via upload
In this task here the user gives input about their weight and height and it returns the bmi index of that person according to the given inputs.

Also the output color changes accordint to the category of the bmi index. For eg: If the person is obese then the color will be red. If they are normal then color will be green and so on.

Here I have used the javascript DOM and event listeners, When the user click on the submit button an event is triggered which takes the values from the field and returns the BMI index.